Break out logic into separate hooks.
