+++ src/features/campaigns/hooks/useActivityOverview.ts
@@ -1,26 +1,14 @@
-import { loadListIfNecessary } from 'core/caching/cacheUtils';
-import { ResolvedFuture } from 'core/caching/futures';
+import { isSameDate } from 'utils/dateUtils';
+import useCallAssignmentActivities from './useCallAssignmentActivities';
 import useEventsFromDateRange from 'features/events/hooks/useEventsFromDateRange';
+import useSurveyActivities from './useSurveyActivities';
+import useTaskActivities from './useTaskActivities';
 import {
-  callAssignmentsLoad,
-  callAssignmentsLoaded,
-  campaignCallAssignmentsLoad,
-  campaignCallAssignmentsLoaded,
-} from 'features/callAssignments/store';
-import {
-  campaignSurveyIdsLoad,
-  campaignSurveyIdsLoaded,
-  surveysLoad,
-  surveysLoaded,
-} from 'features/surveys/store';
-import {
-  campaignTaskIdsLoad,
-  campaignTaskIdsLoaded,
-  tasksLoad,
-  tasksLoaded,
-} from 'features/tasks/store';
-import { getUTCDateWithoutTime, isSameDate } from 'utils/dateUtils';
-import { useApiClient, useAppDispatch, useAppSelector } from 'core/hooks';
+  ErrorFuture,
+  IFuture,
+  LoadingFuture,
+  ResolvedFuture,
+} from 'core/caching/futures';
 import {
   ZetkinCallAssignment,
   ZetkinEvent,
@@ -72,221 +60,43 @@ export type ActivityOverview = {
   tomorrow: CampaignActivity[];
 };
 
-export default function useActivitiyOverview(orgId: number, campId?: number) {
-  const apiClient = useApiClient();
-  const dispatch = useAppDispatch();
-  const tasksSlice = useAppSelector((state) => state.tasks);
-  const surveysSlice = useAppSelector((state) => state.surveys);
-  const callAssignmentsSlice = useAppSelector((state) => state.callAssignments);
-
+export default function useActivitiyOverview(
+  orgId: number,
+  campId?: number
+): IFuture<ActivityOverview> {
   const startOfToday = new Date(new Date().toISOString().slice(0, 10));
   const weekFromNow = new Date(startOfToday);
   weekFromNow.setDate(startOfToday.getDate() + 8);
 
-  const events = useEventsFromDateRange(startOfToday, weekFromNow);
+  const eventActivites = useEventsFromDateRange(startOfToday, weekFromNow);
+  const taskActivitiesFuture = useTaskActivities(orgId, campId);
+  const surveyActivitiesFuture = useSurveyActivities(orgId, campId);
+  const callAssignmentActivitiesFuture = useCallAssignmentActivities(
+    orgId,
+    campId
+  );
+
+  if (
+    callAssignmentActivitiesFuture.isLoading ||
+    surveyActivitiesFuture.isLoading ||
+    taskActivitiesFuture.isLoading
+  ) {
+    return new LoadingFuture();
+  } else if (
+    callAssignmentActivitiesFuture.error ||
+    surveyActivitiesFuture.error ||
+    taskActivitiesFuture.error
+  ) {
+    return new ErrorFuture('Error loading acitvities');
+  }
 
   const activities: CampaignActivity[] = [];
-  activities.push(...events);
-
-  if (campId) {
-    const taskIdsByCampaignId = tasksSlice.taskIdsByCampaignId[campId];
-    const taskIdsFuture = loadListIfNecessary(taskIdsByCampaignId, dispatch, {
-      actionOnLoad: () => campaignTaskIdsLoad(campId),
-      actionOnSuccess: (data) => campaignTaskIdsLoaded([campId, data]),
-      loader: async () => {
-        const tasks = await apiClient.get<ZetkinTask[]>(
-          `/api/orgs/${orgId}/campaigns/${campId}/tasks`
-        );
-        return tasks.map((task) => ({ id: task.id }));
-      },
-    });
-
-    const tasksList = tasksSlice.tasksList;
-    const tasksFuture = loadListIfNecessary(tasksList, dispatch, {
-      actionOnLoad: () => tasksLoad(),
-      actionOnSuccess: (data) => tasksLoaded(data),
-      loader: () =>
-        apiClient.get<ZetkinTask[]>(
-          `/api/orgs/${orgId}/campaigns/${campId}/tasks`
-        ),
-    });
-
-    if (taskIdsFuture.data && tasksFuture.data) {
-      const tasks = tasksFuture.data;
-      const taskIds = taskIdsFuture.data;
-
-      taskIds.forEach(({ id: taskId }) => {
-        const task = tasks.find((item) => item.id === taskId);
-
-        if (task) {
-          activities.push({
-            data: task,
-            kind: ACTIVITIES.TASK,
-            visibleFrom: getUTCDateWithoutTime(task.published || null),
-            visibleUntil: getUTCDateWithoutTime(task.expires || null),
-          });
-        }
-      });
-    }
-
-    const surveyIdsByCampaignId = surveysSlice.surveyIdsByCampaignId[campId];
-    const surveyIdsFuture = loadListIfNecessary(
-      surveyIdsByCampaignId,
-      dispatch,
-      {
-        actionOnLoad: () => campaignSurveyIdsLoad(campId),
-        actionOnSuccess: (data) => campaignSurveyIdsLoaded([campId, data]),
-        loader: async () => {
-          const tasks = await apiClient.get<ZetkinTask[]>(
-            `/api/orgs/${orgId}/campaigns/${campId}/surveys`
-          );
-          return tasks.map((survey) => ({ id: survey.id }));
-        },
-      }
-    );
-
-    const surveyList = surveysSlice.surveyList;
-    const surveysFuture = loadListIfNecessary(surveyList, dispatch, {
-      actionOnLoad: () => surveysLoad(),
-      actionOnSuccess: (data) => surveysLoaded(data),
-      loader: () =>
-        apiClient.get<ZetkinSurvey[]>(
-          `/api/orgs/${orgId}/campaigns/${campId}/surveys`
-        ),
-    });
-
-    if (surveyIdsFuture.data && surveysFuture.data) {
-      const surveys = surveysFuture.data;
-      const surveyIds = surveyIdsFuture.data;
-
-      surveyIds.forEach(({ id: surveyId }) => {
-        const survey = surveys.find((item) => item.id === surveyId);
-
-        if (survey) {
-          activities.push({
-            data: survey,
-            kind: ACTIVITIES.SURVEY,
-            visibleFrom: getUTCDateWithoutTime(survey.published || null),
-            visibleUntil: getUTCDateWithoutTime(survey.expires || null),
-          });
-        }
-      });
-    }
-
-    const callAssignmentIdsByCampaignId =
-      callAssignmentsSlice.callAssignmentIdsByCampaignId[campId];
-    const callAssignmentIdsFuture = loadListIfNecessary(
-      callAssignmentIdsByCampaignId,
-      dispatch,
-      {
-        actionOnLoad: () => campaignCallAssignmentsLoad(campId),
-        actionOnSuccess: (data) =>
-          campaignCallAssignmentsLoaded([campId, data]),
-        loader: async () => {
-          const callAssignments = await apiClient.get<ZetkinCallAssignment[]>(
-            `/api/orgs/${orgId}/campaigns/${campId}/call_assignments`
-          );
-          return callAssignments.map((ca) => ({ id: ca.id }));
-        },
-      }
-    );
-
-    const callAssignmentList = callAssignmentsSlice.assignmentList;
-    const callAssignmentsFuture = loadListIfNecessary(
-      callAssignmentList,
-      dispatch,
-      {
-        actionOnLoad: () => callAssignmentsLoad(),
-        actionOnSuccess: (data) => callAssignmentsLoaded(data),
-        loader: () =>
-          apiClient.get<ZetkinCallAssignment[]>(
-            `/api/orgs/${orgId}/campaigns/${campId}/call_assignments`
-          ),
-      }
-    );
-
-    if (callAssignmentIdsFuture.data && callAssignmentsFuture.data) {
-      const callAssignments = callAssignmentsFuture.data;
-      const callAssignmentIds = callAssignmentIdsFuture.data;
-
-      callAssignmentIds.forEach(({ id: caId }) => {
-        const ca = callAssignments.find((item) => item.id === caId);
-
-        if (ca) {
-          activities.push({
-            data: ca,
-            kind: ACTIVITIES.CALL_ASSIGNMENT,
-            visibleFrom: getUTCDateWithoutTime(ca.start_date),
-            visibleUntil: getUTCDateWithoutTime(ca.end_date),
-          });
-        }
-      });
-    }
-  } else {
-    const tasksList = tasksSlice.tasksList;
-    const allTasksFuture = loadListIfNecessary(tasksList, dispatch, {
-      actionOnLoad: () => tasksLoad(),
-      actionOnSuccess: (data) => tasksLoaded(data),
-      loader: () => apiClient.get<ZetkinTask[]>(`/api/orgs/${orgId}/tasks`),
-    });
-
-    if (allTasksFuture.data) {
-      const allTasks = allTasksFuture.data;
-      allTasks.forEach((task) => {
-        activities.push({
-          data: task,
-          kind: ACTIVITIES.TASK,
-          visibleFrom: getUTCDateWithoutTime(task.published || null),
-          visibleUntil: getUTCDateWithoutTime(task.expires || null),
-        });
-      });
-    }
-
-    const surveyList = surveysSlice.surveyList;
-    const allSurveysFuture = loadListIfNecessary(surveyList, dispatch, {
-      actionOnLoad: () => surveysLoad(),
-      actionOnSuccess: (data) => surveysLoaded(data),
-      loader: () => apiClient.get<ZetkinSurvey[]>(`/api/orgs/${orgId}/surveys`),
-    });
-
-    if (allSurveysFuture.data) {
-      const allSurveys = allSurveysFuture.data;
-      allSurveys.forEach((survey) => {
-        activities.push({
-          data: survey,
-          kind: ACTIVITIES.SURVEY,
-          visibleFrom: getUTCDateWithoutTime(survey.published || null),
-          visibleUntil: getUTCDateWithoutTime(survey.expires || null),
-        });
-      });
-    }
-
-    const callAssignmentList = callAssignmentsSlice.assignmentList;
-    const allCallAssignmentsFuture = loadListIfNecessary(
-      callAssignmentList,
-      dispatch,
-      {
-        actionOnLoad: () => callAssignmentsLoad(),
-        actionOnSuccess: (data) => callAssignmentsLoaded(data),
-        loader: () =>
-          apiClient.get<ZetkinCallAssignment[]>(
-            `/api/orgs/${orgId}/call_assignments`
-          ),
-      }
-    );
-
-    if (allCallAssignmentsFuture.data) {
-      const allCallAssignments = allCallAssignmentsFuture.data;
-      allCallAssignments.forEach((ca) => {
-        activities.push({
-          data: ca,
-          kind: ACTIVITIES.CALL_ASSIGNMENT,
-          visibleFrom: getUTCDateWithoutTime(ca.start_date),
-          visibleUntil: getUTCDateWithoutTime(ca.end_date),
-        });
-      });
-    }
-  }
+  activities.push(
+    ...eventActivites,
+    ...(taskActivitiesFuture.data || []),
+    ...(surveyActivitiesFuture.data || []),
+    ...(callAssignmentActivitiesFuture.data || [])
+  );
 
   const sortedAcitvities = activities.sort((first, second) => {
     if (first.visibleFrom === null) {
